readcmd : more command line syntaxe error handling
In some cases readcmd can cause error. For example a combination of two charaters in '<','>','&','|' will result in error. The error happen when the cmdline struct is free'ed, as `invalid pointer`.

The problem is solved while parsing the command line in `readcmd`, by raising error (as already done) when a "invalid combination" is used. This make the `readcmd` more friendly as it will always output a "valid" command struct.